feat: MovieController working just fine

+++ src/main/java/com/movielog/movielog_api/dtos/request/MovieRequestDTO.java
@@ -1,4 +1,6 @@
 package com.movielog.movielog_api.dtos.request;
 
-public record MovieRequestDTO(String title, Double imdb) {
+import java.math.BigDecimal;
+
+public record MovieRequestDTO(String title, BigDecimal imdb) {
 }
