+++ src/main/java/com/movielog/movielog_api/dtos/response/MovieResponseDTO.java
@@ -1,6 +1,7 @@
 package com.movielog.movielog_api.dtos.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record MovieResponseDTO(Long id, String title, Double imdb, LocalDateTime updatedAt, LocalDateTime createdAt) {
+public record MovieResponseDTO(Long id, String title, BigDecimal imdb, LocalDateTime updatedAt, LocalDateTime createdAt) {
 }
